+++ libft/ft_lstn.c
@@ -6,7 +6,7 @@
 /*   By: kmaitski <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/09 17:03:24 by kmaitski          #+#    #+#             */
-/*   Updated: 2017/03/23 17:01:40 by kmaitski         ###   ########.fr       */
+/*   Updated: 2017/04/12 16:59:07 by kmaitski         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 t_list	*ft_lstn(t_list *head_node, size_t n)
 {
-	size_t	count;
-	t_list	*node;
+	unsigned int	count;
+	t_list			*node;
 
 	if (!head_node)
 		return (NULL);
 	count = 1;
 	node = head_node;
-	while (count++ < n)
+	while ((size_t)count++ < n)
 		node = node->next;
 	return (node);
 }
