+++ libft/ft_strnequ.c
@@ -6,7 +6,7 @@
 /*   By: kmaitski <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/05 18:11:18 by kmaitski          #+#    #+#             */
-/*   Updated: 2017/02/03 13:32:08 by kmaitski         ###   ########.fr       */
+/*   Updated: 2017/04/12 17:29:30 by kmaitski         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strnequ(char const *s1, char const *s2, size_t n)
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	if (!s1 || !s2)
